# backends/common/swagger/swagger.yml
openapi: 3.0.3
info:
  title: Fastify demo API
  description: The official Fastify demo API
  version: 0.1.0
components:
  schemas: {}
paths:
  /api/common/user/register:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  minLength: 1
                  type: string
                password:
                  minLength: 8
                  type: string
              required:
                - name
                - password
        required: true
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                required:
                  - id
                  - name
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/common/user/login:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  minLength: 1
                  type: string
                password:
                  minLength: 1
                  type: string
              required:
                - name
                - password
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                required:
                  - id
                  - name
                  - accessToken
                  - refreshToken
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/common/user/information:
    get:
      tags:
        - User
      parameters:
        - schema:
            minimum: 1
            type: number
          in: query
          name: userId
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  status:
                    type: string
                required:
                  - id
                  - name
                  - status
  /api/common/user/ban:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  minimum: 1
                  type: number
                reason:
                  type: string
              required:
                - userId
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/common/user/block:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  minimum: 1
                  type: number
                targetUserId:
                  minimum: 1
                  type: number
              required:
                - userId
                - targetUserId
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/common/user/destroy_token:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  minimum: 1
                  type: number
                token:
                  minLength: 1
                  type: string
              required:
                - userId
                - token
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/common/user/refresh_token:
    post:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  minLength: 1
                  type: string
              required:
                - refreshToken
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                required:
                  - accessToken
                  - refreshToken
  /api/common/user/password:
    patch:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  minimum: 1
                  type: number
                currentPassword:
                  minLength: 1
                  type: string
                newPassword:
                  minLength: 8
                  type: string
              required:
                - userId
                - currentPassword
                - newPassword
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/common/user/delete:
    delete:
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  minimum: 1
                  type: number
                reason:
                  type: string
              required:
                - userId
        required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/common/auth/users:
    get:
      tags:
        - Auth
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: number
                    name:
                      type: string
                    password:
                      type: string
                    salt:
                      type: string
                  required:
                    - id
                    - name
                    - password
                    - salt
  /api/common/auth/login:
    post:
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  minLength: 1
                  type: string
                password:
                  minLength: 1
                  type: string
              required:
                - name
                - password
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                required:
                  - id
                  - name
                  - accessToken
                  - refreshToken
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message
  /api/common/auth/me:
    post:
      tags:
        - Auth
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                  name:
                    type: string
                required:
                  - id
                  - name
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                  - message